+++ libfaac/blockswitch.h
@@ -39,7 +39,7 @@ typedef struct {
 	int sizeS;
 
 	/* Previous input samples */
-	double *prevSamples;
+	float *prevSamples;
 
 	int block_type;
 
@@ -47,11 +47,11 @@ typedef struct {
 } PsyInfo;
 
 typedef struct {
-	double sampleRate;
+	float sampleRate;
 
 	/* Hann window */
-	double *hannWindow;
-	double *hannWindowS;
+	float *hannWindow;
+	float *hannWindowS;
 
         void *data;
 } GlobalPsyInfo;
@@ -67,9 +67,9 @@ void (*PsyEnd) (GlobalPsyInfo *gpsyInfo, PsyInfo *psyInfo,
 void (*PsyCalculate) (ChannelInfo *channelInfo, GlobalPsyInfo *gpsyInfo,
 		PsyInfo *psyInfo, int *cb_width_long, int num_cb_long,
 		int *cb_width_short, int num_cb_short,
-		unsigned int numChannels, double quality);
+		unsigned int numChannels, float quality);
 void (*PsyBufferUpdate) ( FFT_Tables *fft_tables, GlobalPsyInfo * gpsyInfo, PsyInfo * psyInfo,
-		double *newSamples, unsigned int bandwidth,
+		float *newSamples, unsigned int bandwidth,
 		int *cb_width_short, int num_cb_short);
 void (*BlockSwitch) (CoderInfo *coderInfo, PsyInfo *psyInfo,
 		unsigned int numChannels);
