+++ libfaac/filtbank.c
@@ -44,10 +44,10 @@ Copyright(c)1996.
 #define  TWOPI       2*M_PI
 
 
-static void		CalculateKBDWindow	( double* win, double alpha, int length );
-static double	Izero				( double x);
-static void		MDCT				( FFT_Tables *fft_tables, double *data, int N );
-static void		IMDCT				( FFT_Tables *fft_tables, double *data, int N );
+static void		CalculateKBDWindow	( float* win, float alpha, int length );
+static float	Izero				( float x);
+static void		MDCT				( FFT_Tables *fft_tables, float *data, int N );
+static void		IMDCT				( FFT_Tables *fft_tables, float *data, int N );
 
 
 
@@ -56,15 +56,15 @@ void FilterBankInit(faacEncStruct* hEncoder)
     unsigned int i, channel;
 
     for (channel = 0; channel < hEncoder->numChannels; channel++) {
-        hEncoder->freqBuff[channel] = (double*)AllocMemory(2*FRAME_LEN*sizeof(double));
-        hEncoder->overlapBuff[channel] = (double*)AllocMemory(FRAME_LEN*sizeof(double));
-        SetMemory(hEncoder->overlapBuff[channel], 0, FRAME_LEN*sizeof(double));
+        hEncoder->freqBuff[channel] = (float*)AllocMemory(2*FRAME_LEN*sizeof(float));
+        hEncoder->overlapBuff[channel] = (float*)AllocMemory(FRAME_LEN*sizeof(float));
+        SetMemory(hEncoder->overlapBuff[channel], 0, FRAME_LEN*sizeof(float));
     }
 
-    hEncoder->sin_window_long = (double*)AllocMemory(BLOCK_LEN_LONG*sizeof(double));
-    hEncoder->sin_window_short = (double*)AllocMemory(BLOCK_LEN_SHORT*sizeof(double));
-    hEncoder->kbd_window_long = (double*)AllocMemory(BLOCK_LEN_LONG*sizeof(double));
-    hEncoder->kbd_window_short = (double*)AllocMemory(BLOCK_LEN_SHORT*sizeof(double));
+    hEncoder->sin_window_long = (float*)AllocMemory(BLOCK_LEN_LONG*sizeof(float));
+    hEncoder->sin_window_short = (float*)AllocMemory(BLOCK_LEN_SHORT*sizeof(float));
+    hEncoder->kbd_window_long = (float*)AllocMemory(BLOCK_LEN_LONG*sizeof(float));
+    hEncoder->kbd_window_short = (float*)AllocMemory(BLOCK_LEN_SHORT*sizeof(float));
 
     for( i=0; i<BLOCK_LEN_LONG; i++ )
         hEncoder->sin_window_long[i] = sin((M_PI/(2*BLOCK_LEN_LONG)) * (i + 0.5));
@@ -92,26 +92,26 @@ void FilterBankEnd(faacEncStruct* hEncoder)
 
 void FilterBank(faacEncStruct* hEncoder,
                 CoderInfo *coderInfo,
-                double *p_in_data,
-                double *p_out_mdct,
-                double *p_overlap,
+                float *p_in_data,
+                float *p_out_mdct,
+                float *p_overlap,
                 int overlap_select)
 {
-    double *p_o_buf, *first_window, *second_window;
-    double *transf_buf;
+    float *p_o_buf, *first_window, *second_window;
+    float *transf_buf;
     int k, i;
     int block_type = coderInfo->block_type;
 
-    transf_buf = (double*)AllocMemory(2*BLOCK_LEN_LONG*sizeof(double));
+    transf_buf = (float*)AllocMemory(2*BLOCK_LEN_LONG*sizeof(float));
 
     /* create / shift old values */
     /* We use p_overlap here as buffer holding the last frame time signal*/
     if(overlap_select != MNON_OVERLAPPED) {
-        memcpy(transf_buf, p_overlap, FRAME_LEN*sizeof(double));
-        memcpy(transf_buf+BLOCK_LEN_LONG, p_in_data, FRAME_LEN*sizeof(double));
-        memcpy(p_overlap, p_in_data, FRAME_LEN*sizeof(double));
+        memcpy(transf_buf, p_overlap, FRAME_LEN*sizeof(float));
+        memcpy(transf_buf+BLOCK_LEN_LONG, p_in_data, FRAME_LEN*sizeof(float));
+        memcpy(p_overlap, p_in_data, FRAME_LEN*sizeof(float));
     } else {
-        memcpy(transf_buf, p_in_data, 2*FRAME_LEN*sizeof(double));
+        memcpy(transf_buf, p_in_data, 2*FRAME_LEN*sizeof(float));
     }
 
     /*  Window shape processing */
@@ -169,18 +169,18 @@ void FilterBank(faacEncStruct* hEncoder,
     case LONG_SHORT_WINDOW :
         for ( i = 0 ; i < BLOCK_LEN_LONG ; i++)
             p_out_mdct[i] = p_o_buf[i] * first_window[i];
-        memcpy(p_out_mdct+BLOCK_LEN_LONG,p_o_buf+BLOCK_LEN_LONG,NFLAT_LS*sizeof(double));
+        memcpy(p_out_mdct+BLOCK_LEN_LONG,p_o_buf+BLOCK_LEN_LONG,NFLAT_LS*sizeof(float));
         for ( i = 0 ; i < BLOCK_LEN_SHORT ; i++)
             p_out_mdct[i+BLOCK_LEN_LONG+NFLAT_LS] = p_o_buf[i+BLOCK_LEN_LONG+NFLAT_LS] * second_window[BLOCK_LEN_SHORT-i-1];
-        SetMemory(p_out_mdct+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(double));
+        SetMemory(p_out_mdct+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(float));
         MDCT( &hEncoder->fft_tables, p_out_mdct, 2*BLOCK_LEN_LONG );
         break;
 
     case SHORT_LONG_WINDOW :
-        SetMemory(p_out_mdct,0,NFLAT_LS*sizeof(double));
+        SetMemory(p_out_mdct,0,NFLAT_LS*sizeof(float));
         for ( i = 0 ; i < BLOCK_LEN_SHORT ; i++)
             p_out_mdct[i+NFLAT_LS] = p_o_buf[i+NFLAT_LS] * first_window[i];
-        memcpy(p_out_mdct+NFLAT_LS+BLOCK_LEN_SHORT,p_o_buf+NFLAT_LS+BLOCK_LEN_SHORT,NFLAT_LS*sizeof(double));
+        memcpy(p_out_mdct+NFLAT_LS+BLOCK_LEN_SHORT,p_o_buf+NFLAT_LS+BLOCK_LEN_SHORT,NFLAT_LS*sizeof(float));
         for ( i = 0 ; i < BLOCK_LEN_LONG ; i++)
             p_out_mdct[i+BLOCK_LEN_LONG] = p_o_buf[i+BLOCK_LEN_LONG] * second_window[BLOCK_LEN_LONG-i-1];
         MDCT( &hEncoder->fft_tables, p_out_mdct, 2*BLOCK_LEN_LONG );
@@ -206,20 +206,20 @@ void FilterBank(faacEncStruct* hEncoder,
 
 void IFilterBank(faacEncStruct* hEncoder,
                  CoderInfo *coderInfo,
-                 double *p_in_data,
-                 double *p_out_data,
-                 double *p_overlap,
+                 float *p_in_data,
+                 float *p_out_data,
+                 float *p_overlap,
                  int overlap_select)
 {
-    double *o_buf, *transf_buf, *overlap_buf;
-    double *first_window, *second_window;
+    float *o_buf, *transf_buf, *overlap_buf;
+    float *first_window, *second_window;
 
-    double  *fp;
+    float  *fp;
     int k, i;
     int block_type = coderInfo->block_type;
 
-    transf_buf = (double*)AllocMemory(2*BLOCK_LEN_LONG*sizeof(double));
-    overlap_buf = (double*)AllocMemory(2*BLOCK_LEN_LONG*sizeof(double));
+    transf_buf = (float*)AllocMemory(2*BLOCK_LEN_LONG*sizeof(float));
+    overlap_buf = (float*)AllocMemory(2*BLOCK_LEN_LONG*sizeof(float));
 
     /*  Window shape processing */
     if (overlap_select != MNON_OVERLAPPED) {
@@ -259,13 +259,13 @@ void IFilterBank(faacEncStruct* hEncoder,
     }
 
     /* Assemble overlap buffer */
-    memcpy(overlap_buf,p_overlap,BLOCK_LEN_LONG*sizeof(double));
+    memcpy(overlap_buf,p_overlap,BLOCK_LEN_LONG*sizeof(float));
     o_buf = overlap_buf;
 
     /* Separate action for each Block Type */
     switch( block_type ) {
     case ONLY_LONG_WINDOW :
-        memcpy(transf_buf, p_in_data,BLOCK_LEN_LONG*sizeof(double));
+        memcpy(transf_buf, p_in_data,BLOCK_LEN_LONG*sizeof(float));
         IMDCT( &hEncoder->fft_tables, transf_buf, 2*BLOCK_LEN_LONG );
         for ( i = 0 ; i < BLOCK_LEN_LONG ; i++)
             transf_buf[i] *= first_window[i];
@@ -281,37 +281,37 @@ void IFilterBank(faacEncStruct* hEncoder,
         break;
 
     case LONG_SHORT_WINDOW :
-        memcpy(transf_buf, p_in_data,BLOCK_LEN_LONG*sizeof(double));
+        memcpy(transf_buf, p_in_data,BLOCK_LEN_LONG*sizeof(float));
         IMDCT( &hEncoder->fft_tables, transf_buf, 2*BLOCK_LEN_LONG );
         for ( i = 0 ; i < BLOCK_LEN_LONG ; i++)
             transf_buf[i] *= first_window[i];
         if (overlap_select != MNON_OVERLAPPED) {
             for ( i = 0 ; i < BLOCK_LEN_LONG; i++ )
                 o_buf[i] += transf_buf[i];
-            memcpy(o_buf+BLOCK_LEN_LONG,transf_buf+BLOCK_LEN_LONG,NFLAT_LS*sizeof(double));
+            memcpy(o_buf+BLOCK_LEN_LONG,transf_buf+BLOCK_LEN_LONG,NFLAT_LS*sizeof(float));
             for ( i = 0 ; i < BLOCK_LEN_SHORT ; i++)
                 o_buf[i+BLOCK_LEN_LONG+NFLAT_LS] = transf_buf[i+BLOCK_LEN_LONG+NFLAT_LS] * second_window[BLOCK_LEN_SHORT-i-1];
-            SetMemory(o_buf+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(double));
+            SetMemory(o_buf+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(float));
         } else { /* overlap_select == NON_OVERLAPPED */
             for ( i = 0 ; i < BLOCK_LEN_SHORT ; i++)
                 transf_buf[i+BLOCK_LEN_LONG+NFLAT_LS] *= second_window[BLOCK_LEN_SHORT-i-1];
-            SetMemory(transf_buf+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(double));
+            SetMemory(transf_buf+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(float));
         }
         break;
 
     case SHORT_LONG_WINDOW :
-        memcpy(transf_buf, p_in_data,BLOCK_LEN_LONG*sizeof(double));
+        memcpy(transf_buf, p_in_data,BLOCK_LEN_LONG*sizeof(float));
         IMDCT( &hEncoder->fft_tables, transf_buf, 2*BLOCK_LEN_LONG );
         for ( i = 0 ; i < BLOCK_LEN_SHORT ; i++)
             transf_buf[i+NFLAT_LS] *= first_window[i];
         if (overlap_select != MNON_OVERLAPPED) {
             for ( i = 0 ; i < BLOCK_LEN_SHORT; i++ )
                 o_buf[i+NFLAT_LS] += transf_buf[i+NFLAT_LS];
-            memcpy(o_buf+BLOCK_LEN_SHORT+NFLAT_LS,transf_buf+BLOCK_LEN_SHORT+NFLAT_LS,NFLAT_LS*sizeof(double));
+            memcpy(o_buf+BLOCK_LEN_SHORT+NFLAT_LS,transf_buf+BLOCK_LEN_SHORT+NFLAT_LS,NFLAT_LS*sizeof(float));
             for ( i = 0 ; i < BLOCK_LEN_LONG ; i++)
                 o_buf[i+BLOCK_LEN_LONG] = transf_buf[i+BLOCK_LEN_LONG] * second_window[BLOCK_LEN_LONG-i-1];
         } else { /* overlap_select == NON_OVERLAPPED */
-            SetMemory(transf_buf,0,NFLAT_LS*sizeof(double));
+            SetMemory(transf_buf,0,NFLAT_LS*sizeof(float));
             for ( i = 0 ; i < BLOCK_LEN_LONG ; i++)
                 transf_buf[i+BLOCK_LEN_LONG] *= second_window[BLOCK_LEN_LONG-i-1];
         }
@@ -324,7 +324,7 @@ void IFilterBank(faacEncStruct* hEncoder,
             fp = transf_buf;
         }
         for ( k=0; k < MAX_SHORT_WINDOWS; k++ ) {
-            memcpy(transf_buf,p_in_data,BLOCK_LEN_SHORT*sizeof(double));
+            memcpy(transf_buf,p_in_data,BLOCK_LEN_SHORT*sizeof(float));
             IMDCT( &hEncoder->fft_tables, transf_buf, 2*BLOCK_LEN_SHORT );
             p_in_data += BLOCK_LEN_SHORT;
             if (overlap_select != MNON_OVERLAPPED) {
@@ -343,23 +343,23 @@ void IFilterBank(faacEncStruct* hEncoder,
             }
             first_window = second_window;
         }
-        SetMemory(o_buf+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(double));
+        SetMemory(o_buf+BLOCK_LEN_LONG+NFLAT_LS+BLOCK_LEN_SHORT,0,NFLAT_LS*sizeof(float));
         break;
     }
 
     if (overlap_select != MNON_OVERLAPPED)
-        memcpy(p_out_data,o_buf,BLOCK_LEN_LONG*sizeof(double));
+        memcpy(p_out_data,o_buf,BLOCK_LEN_LONG*sizeof(float));
     else  /* overlap_select == NON_OVERLAPPED */
-        memcpy(p_out_data,transf_buf,2*BLOCK_LEN_LONG*sizeof(double));
+        memcpy(p_out_data,transf_buf,2*BLOCK_LEN_LONG*sizeof(float));
 
     /* save unused output data */
-    memcpy(p_overlap,o_buf+BLOCK_LEN_LONG,BLOCK_LEN_LONG*sizeof(double));
+    memcpy(p_overlap,o_buf+BLOCK_LEN_LONG,BLOCK_LEN_LONG*sizeof(float));
 
     if (overlap_buf) FreeMemory(overlap_buf);
     if (transf_buf) FreeMemory(transf_buf);
 }
 
-void specFilter(double *freqBuff,
+void specFilter(float *freqBuff,
                 int sampleRate,
                 int lowpassFreq,
                 int specLen
@@ -371,19 +371,19 @@ void specFilter(double *freqBuff,
     lowpass = (lowpassFreq * specLen) / (sampleRate>>1) + 1;
     xlowpass = (lowpass < specLen) ? lowpass : specLen ;
 
-    SetMemory(freqBuff+xlowpass,0,(specLen-xlowpass)*sizeof(double));
+    SetMemory(freqBuff+xlowpass,0,(specLen-xlowpass)*sizeof(float));
 }
 
-static double Izero(double x)
+static float Izero(float x)
 {
-    const double IzeroEPSILON = 1E-41;  /* Max error acceptable in Izero */
-    double sum, u, halfx, temp;
+    const float IzeroEPSILON = 1E-41;  /* Max error acceptable in Izero */
+    float sum, u, halfx, temp;
     int n;
 
     sum = u = n = 1;
     halfx = x/2.0;
     do {
-        temp = halfx/(double)n;
+        temp = halfx/(float)n;
         n += 1;
         temp *= temp;
         u *= temp;
@@ -393,19 +393,19 @@ static double Izero(double x)
     return(sum);
 }
 
-static void CalculateKBDWindow(double* win, double alpha, int length)
+static void CalculateKBDWindow(float* win, float alpha, int length)
 {
     int i;
-    double IBeta;
-    double tmp;
-    double sum = 0.0;
+    float IBeta;
+    float tmp;
+    float sum = 0.0;
 
     alpha *= M_PI;
     IBeta = 1.0/Izero(alpha);
 
     /* calculate lower half of Kaiser Bessel window */
     for(i=0; i<(length>>1); i++) {
-        tmp = 4.0*(double)i/(double)length - 1.0;
+        tmp = 4.0*(float)i/(float)length - 1.0;
         win[i] = Izero(alpha*sqrt(1.0-tmp*tmp))*IBeta;
         sum += win[i];
     }
@@ -420,16 +420,16 @@ static void CalculateKBDWindow(double* win, double alpha, int length)
     }
 }
 
-static void MDCT( FFT_Tables *fft_tables, double *data, int N )
+static void MDCT( FFT_Tables *fft_tables, float *data, int N )
 {
-    double *xi, *xr;
-    double tempr, tempi, c, s, cold, cfreq, sfreq; /* temps for pre and post twiddle */
-    double freq = TWOPI / N;
-    double cosfreq8, sinfreq8;
+    float *xi, *xr;
+    float tempr, tempi, c, s, cold, cfreq, sfreq; /* temps for pre and post twiddle */
+    float freq = TWOPI / N;
+    float cosfreq8, sinfreq8;
     int i, n;
 
-    xi = (double*)AllocMemory((N >> 2)*sizeof(double));
-    xr = (double*)AllocMemory((N >> 2)*sizeof(double));
+    xi = (float*)AllocMemory((N >> 2)*sizeof(float));
+    xr = (float*)AllocMemory((N >> 2)*sizeof(float));
 
     /* prepare for recurrence relation in pre-twiddle */
     cfreq = cos (freq);
@@ -499,16 +499,16 @@ static void MDCT( FFT_Tables *fft_tables, double *data, int N )
     if (xi) FreeMemory(xi);
 }
 
-static void IMDCT( FFT_Tables *fft_tables, double *data, int N)
+static void IMDCT( FFT_Tables *fft_tables, float *data, int N)
 {
-    double *xi, *xr;
-    double tempr, tempi, c, s, cold, cfreq, sfreq; /* temps for pre and post twiddle */
-    double freq = 2.0 * M_PI / N;
-    double fac, cosfreq8, sinfreq8;
+    float *xi, *xr;
+    float tempr, tempi, c, s, cold, cfreq, sfreq; /* temps for pre and post twiddle */
+    float freq = 2.0 * M_PI / N;
+    float fac, cosfreq8, sinfreq8;
     int i;
 
-    xi = (double*)AllocMemory((N >> 2)*sizeof(double));
-    xr = (double*)AllocMemory((N >> 2)*sizeof(double));
+    xi = (float*)AllocMemory((N >> 2)*sizeof(float));
+    xr = (float*)AllocMemory((N >> 2)*sizeof(float));
 
     /* Choosing to allocate 2/N factor to Inverse Xform! */
     fac = 2. / N; /* remaining 2/N from 4/N IFFT factor */
